+++ fhir-server/src/main/java/au/csiro/pathling/fhirpath/function/subsumes/SubsumesFunction.java
@@ -12,8 +12,8 @@ import au.csiro.pathling.fhirpath.element.BooleanPath;
 import au.csiro.pathling.fhirpath.element.ElementPath;
 import au.csiro.pathling.fhirpath.function.NamedFunction;
 import au.csiro.pathling.fhirpath.function.NamedFunctionInput;
-import au.csiro.pathling.fhirpath.function.subsumes.encoding.IdAndBoolean;
-import au.csiro.pathling.fhirpath.function.subsumes.encoding.IdAndCodingSets;
+import au.csiro.pathling.fhirpath.encoding.IdAndBoolean;
+import au.csiro.pathling.fhirpath.encoding.IdAndCodingSets;
 import au.csiro.pathling.fhirpath.literal.CodingLiteralPath;
 import au.csiro.pathling.fhirpath.parser.ParserContext;
 import java.util.Optional;
